+++ BitReal/BitReal/Components/Habit.swift
@@ -4,6 +4,7 @@
 //
 //  Created by Don Nguyen on 3/1/23.
 //
+// DEPRECATED
 
 import SwiftUI
